Split journal basic api and entity api endpoints, added handlers
