+++ sample.env
@@ -37,7 +37,7 @@ export BUGOUT_BOT_INSTALLATION_TOKEN="<installation user token from brood>"
 export BUGOUT_BOT_INSTALLATION_TOKEN_HEADER="<bugout installation token header>"
 export BUGOUT_DRONES_TOKEN="<auth internal drones token>"
 export BUGOUT_DRONES_TOKEN_HEADER="<auth internal drones token header>"
-export SPIRE_OPENAPI_LIST="journals,humbug,preferences,public,go"
+export SPIRE_OPENAPI_LIST="journals,collections,humbug,preferences,public,go"
 export BUGOUT_REDIS_URL="http://127.0.0.1:6379"
 export BUGOUT_REDIS_PASSWORD="mypassword"
 export REDIS_REPORTS_QUEUE="<redis key to humbug reports queue>"
