+++ spire/journal/version.py
@@ -1,5 +1,6 @@
 """
-Journals module of Spire library and API version.
+Journal and collection modules of Spire library and API version.
 """
 
 SPIRE_JOURNALS_VERSION = "0.1.2"
+SPIRE_COLLECTIONS_VERSION = "0.0.7"
